python/mcp-mariadb/pyproject.toml
[project]
name = "mcp-mariadbr"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncmy>=0.2.10",
    "fastmcp[cli]==2.2.8",
    "fastmcp[websockets]==2.12.1",
    "google-genai>=1.15.0",
    # "google-generativeai>=0.8.5",
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "asyncmy>=0.2.10",
    "tokenizers==0.21.2",
    "numpy",
]
